system regular upate

+++ src/TP031_Update_Idx_Hist.py
@@ -7,36 +7,41 @@ Created on Tue Apr 10 06:19:07 2018
 
 import os,time
 import pandas as pd
-import fdmt_data as fdt
-import fdmt_date as fda
+import fdmt_data
+import fdmt_date
 
 def main():
-    data_path=fdt.data_path
+    data_path=fdmt_data.data_path
+    magic_box=fdmt_data.magic_box
 
-    magic_box=fdt.magic_box
+    para_date_start=fdmt_date.current_date_str
+    para_date_end=fdmt_date.current_date_str
 
-    para_date_start=fda.current_date_str
-    para_date_end=fda.current_date_str
-
-    if fda.current_date_str>=fda.date_add(fda.current_month_str,5):
-        para_date_start=fda.current_month_str
-        para_date_end=fda.date_add(fda.current_month_str,1,"MM")
+    if fdmt_date.current_date_str>=fdmt_date.date_add(fdmt_date.current_month_str,5):
+        para_date_start=fdmt_date.current_month_str
+        para_date_end=fdmt_date.date_add(fdmt_date.current_month_str,1,"MM")
     else:
-        para_date_start=fda.date_add(fda.current_month_str,-1,"MM")
-        para_date_end=fda.date_add(fda.current_month_str,1,"MM")
+        para_date_start=fdmt_date.date_add(fdmt_date.current_month_str,-1,"MM")
+        para_date_end=fdmt_date.date_add(fdmt_date.current_month_str,1,"MM")
 
-    #para_date_start='2018-04-01'
-    #para_date_end='2018-06-01'
-    
-    para_mon_start=fda.date_format(para_date_start,"yyyymm")
-    para_mon_end=fda.date_format(para_date_end,"yyyymm")
+    #Be careful!!!
+    inital_flag=False
 
-    fdt.pgs_execute_query("TRUNCATE TABLE MAIN.TMP_IDX_HIST")
+    if inital_flag==True:
+        para_date_start='2014-01-01'
+        para_date_end=fdmt_date.current_date_str
+
+    para_mon_start=fdmt_date.date_format(para_date_start,"yyyymm")
+    para_mon_end=fdmt_date.date_format(para_date_end,"yyyymm")
+
+
+    if inital_flag==False:
+        fdt.pgs_execute_query("TRUNCATE TABLE MAIN.TMP_IDX_HIST")
 
     hist_list=[]
-    for parent,dirnames,filenames in os.walk(data_path+"idx_hist/"):
+    for parent,dirnames,filenames in os.walk(data_path+"idx_pkl/"):
         for filename in filenames:
-            if (filename[-4:].lower()=='.txt')&(filename[-17:-11]>=para_mon_start)&(filename[-17:-11]<=para_mon_end):
+            if (filename[-4:].lower()=='.pkl')&(filename[-17:-11]>=para_mon_start)&(filename[-17:-11]<=para_mon_end):
                 hist_list.append(parent+filename)
 
     for para in hist_list:
@@ -47,6 +52,9 @@ def main():
         para_df2["utime"]=fda.current_time_str
 
         para_df2.to_sql("tmp_idx_hist",magic_box,if_exists="append",schema="main",index=False)
+
+    if inital_flag==True:
+        fdt.pgs_execute_query("CREATE TABLE MAIN.F_IDX_HIST AS  WHERE TDATE>=(SELECT MIN(TDATE) FROM MAIN.TMP_IDX_HIST)")
 
     fdt.pgs_execute_query("DELETE FROM MAIN.F_IDX_HIST WHERE TDATE>=(SELECT MIN(TDATE) FROM MAIN.TMP_IDX_HIST)")
     fdt.pgs_execute_query("INSERT INTO MAIN.F_IDX_HIST SELECT * FROM MAIN.TMP_IDX_HIST")
